perf: WALLET-1381 — build the replica store without the background's sagas
createMainStoreReplica used createStore from @background/redux, the same
factory the background singleton uses, which installs the saga middleware
and runs rootSaga. So every frontend page linked and ran the whole
background saga graph — including check-casper2-network-saga, which
imports HttpHandler and RpcClient, and so put casper-js-sdk on the
startup path of all five page entries through the store they render from.

Nothing in a replica can trigger a saga. Every frontend dispatch goes
through dispatchToMainStore (runtime.sendMessage) to the background, and
there is no useDispatch anywhere in src/ — no watcher in a replica could
ever fire. The store is also rebuilt on each popupStateUpdated broadcast,
so every broadcast started a fresh set of watcher tasks and dropped the
previous ones.

Had an action ever reached a replica, the sagas would have run in a page
context: openExportKeysWindow opens the private-key export window (once
per open replica, racing the tracked window id), resetVault calls
storage.local.clear(), the vault sagas broadcast SDK events to every dapp
tab, and the two timer sagas write the auto-lock deadlines the background
resumes from.

The replica now has its own module and builds a plain configureStore with
the same middleware options minus the saga middleware. createStore keeps
running rootSaga for the background, which is where it belongs.
