+++ src/apps/connect-to-app/index.tsx
@@ -9,7 +9,7 @@ import { isSafariBuild, setCSPForSafari } from '@src/utils';
 
 import { AppRouter } from '@connect-to-app/app-router';
 
-import { createMainStoreReplica } from '@background/redux/get-main-store';
+import { createMainStoreReplica } from '@background/redux/create-main-store-replica';
 import { themeModeSettingChanged } from '@background/redux/settings/actions';
 import { selectThemeModeSetting } from '@background/redux/settings/selectors';
 import { ThemeMode } from '@background/redux/settings/types';
