+++ src/apps/onboarding/index.tsx
@@ -9,7 +9,7 @@ import { setCSPForSafari } from '@src/utils';
 
 import { AppRouter } from '@onboarding/app-router';
 
-import { createMainStoreReplica } from '@background/redux/get-main-store';
+import { createMainStoreReplica } from '@background/redux/create-main-store-replica';
 import { PopupState } from '@background/redux/types';
 import { onboardingAppInit } from '@background/redux/windowManagement/actions';
